removed duplicate definitions
