+++ server/shared/types/terms/samplelst.ts
@@ -1,4 +1,5 @@
-import { Term, TermWrapper, BaseQ } from '../termdb.ts'
+import { Term, BaseQ } from '../term.ts'
+import { TermWrapper } from './tw.js'
 import { TermSettingInstance } from '../termsetting.ts'
 
 /*
